Move fetchsm into proxy
