+++ include/proxy/PluginHttpConnect.h
@@ -1,6 +1,6 @@
 /** @file
 
-  Stub file for linking libinknet.a from unit tests
+  A brief file description
 
   @section license License
 
@@ -21,45 +21,8 @@
   limitations under the License.
  */
 
-#include "tscore/Version.h"
+#pragma once
 
-AppVersionInfo appVersionInfo;
+#include "proxy/PluginVC.h"
 
-#include "proxy/FetchSM.h"
-ClassAllocator<FetchSM> FetchSMAllocator("unusedFetchSMAllocator");
-void
-FetchSM::ext_launch()
-{
-}
-void
-FetchSM::ext_destroy()
-{
-}
-ssize_t
-FetchSM::ext_read_data(char *, unsigned long)
-{
-  return 0;
-}
-void
-FetchSM::ext_add_header(char const *, int, char const *, int)
-{
-}
-void
-FetchSM::ext_write_data(void const *, unsigned long)
-{
-}
-void *
-FetchSM::ext_get_user_data()
-{
-  return nullptr;
-}
-void
-FetchSM::ext_set_user_data(void *)
-{
-}
-void
-FetchSM::ext_init(Continuation *, char const *, char const *, char const *, sockaddr const *, int)
-{
-}
-
-ChunkedHandler::ChunkedHandler() {}
+PluginVC *PluginHttpConnectInternal(TSHttpConnectOptions *options);
