+++ src/iocore/cache/unit_tests/stub.cc
@@ -78,7 +78,7 @@ TSIOBufferReaderConsume(TSIOBufferReader readerp, int64_t nbytes)
 }
 } // namespace tsapi::c
 
-#include "api/FetchSM.h"
+#include "proxy/FetchSM.h"
 ClassAllocator<FetchSM> FetchSMAllocator("unusedFetchSMAllocator");
 void
 FetchSM::ext_launch()
